pyproject.toml
[tool.poetry]
name = "store"
version = "0.1.0"
description = ""
authors = [""]
readme = "readme.rst"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
djangorestframework = "^3.14.0"
factory-boy = "^3.3.0"
django-redis = "^5.4.0"
dj-database-url = "^2.1.0"
sentry-sdk = "^1.39.2"
redis = {extras = ["hiredis"], version = "^5.0.1"}
celery = {extras = ["redis"], version = "^5.3.6"}
uwsgi = "^2.0.23"
psycopg2-binary = "^2.9.9"
requests = "^2.31.0"
django-push-notifications = "^3.0.0"
django-money = "^3.0.0"
drf-spectacular = "^0.27.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
